# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Artifex API",
        "description": "二维游戏美术协作与 AI 资产生成平台 API",
        "version": "1.4.0",
        "contact": { "name": "Artifex Team" }
    },
    "servers": [
        { "url": "http://127.0.0.1:3000", "description": "本地开发" }
    ],
...
    "paths": {
...
        "/api/image-proxy": {
            "post": {
                "tags": ["image"],
                "summary": "图片生成（文生图 / 图生图）",
                "description": "支持 tencent / alibaba / jimeng / sdwebui / free / mock 等 provider",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["prompt"],
                                "properties": {
                                    "prompt": { "type": "string", "description": "生成提示词" },
                                    "size": { "type": "string", "description": "尺寸，如 1024x1024" },
                                    "provider": { "type": "string", "enum": ["tencent", "alibaba", "jimeng", "sdwebui", "free", "mock"], "default": "tencent" },
                                    "mode": { "type": "string", "enum": ["text2img", "img2img"], "default": "text2img" },
                                    "image": { "type": "string", "description": "图生图的 base64 / data URL" },
                                    "strength": { "type": "number", "default": 0.7 },
                                    "imageModel": { "type": "string" },
                                    "styleReferenceImage": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "生成成功，返回 image_url" },
                    "400": { "description": "参数错误" },
                    "502": { "description": "上游 API 失败" }
                }
            }
        },
        "/api/proxy-image": {
            "get": {
                "tags": ["image"],
                "summary": "代理拉取远程图片",
                "parameters": [{ "name": "url", "in": "query", "required": true, "schema": { "type": "string" } }],
                "responses": { "200": { "description": "图片内容" } }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            }
        }
    },
    "security": [{ "cookieAuth": [] }]
}